% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti_item.R
\name{qti_item}
\alias{qti_item}
\title{The qti_item class}
\usage{
qti_item(id = NULL, title = NULL, type = c("choice"), prompt = NULL,
  options = NULL, key = NULL, xml = NULL)
}
\arguments{
\item{id}{Item ID as an integer or string.}

\item{title}{Item title, as a string, with max 140 characters.}

\item{type}{Item type, only "choice" is currently supported.}

\item{prompt}{Stem text, as a single string.}

\item{options}{Option text, as vector of strings, one per option.}

\item{key}{Numeric vector of 0/1 response scores, one per option.}

\item{xml}{Optional XML representation of item, as character. If not
supplied, this is created automatically from the remaining arguments.}
}
\value{
\code{qti_item} returns an object of class
\dQuote{\code{qti_item}}, which is a list containing named elements for
the item \code{id}, \code{title}, and \code{type}, as supplied via
arguments, the \code{prompt} and \code{options} text with formatting and
resources stripped, and the QTI XML representation of the item.
}
\description{
These functions build and manipulate qti_item objects, for analysis and
for reading and writing to QTI XML.
}
\examples{

item <- qti_item(
  id = 1584,
  title = "Excellent Example Item",
  type = "choice",
  prompt = "Is this a really good question?",
  options = c("Yes", "No", "Maybe"),
  key = c(0, 0, 1)
)
item

}
\keyword{methods}
