% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti_manifest.R
\name{qti_manifest}
\alias{qti_manifest}
\title{The qti_manifest class}
\usage{
qti_manifest(x, id = NULL, title = NULL, xml = NULL)
}
\arguments{
\item{x}{One or more items as \dQuote{\code{qti_item}} objects.}

\item{id}{Optional identifier.}

\item{title}{Assessment title, as a string, with max 140 characters.}

\item{xml}{Optional XML representation of manifest, as character. If not
supplied, this is created automatically from the remaining arguments.}
}
\value{
\code{qti_manifest} returns an object of class
\dQuote{\code{qti_manifest}}, which is a list containing named elements
for the assessment \code{id} and \code{title}, as supplied via
arguments, the items in \code{x} as a list of \dQuote{\code{qti_item}},
and the QTI XML representation of the manifest.
}
\description{
These functions build and manipulate qti_manifest objects, for analysis
and for reading and writing to QTI XML.
}
\examples{

item <- qti_item(
  id = "item-1",
  title = "Example Item",
  type = "choice",
  prompt = "What does this image tell you? <img src='life.png' />",
  options = c("Everything", "Something",
    "Nothing, but look at this code:<br/><pre>lm(life ~ R)</pre>"),
  key = c(1, 1, 0)
)
man <- qti_manifest(
  x = item,
  id = "exam-1",
  title = "Example Exam"
)

}
\keyword{methods}
