% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_qti.R
\name{write_qti}
\alias{write_qti}
\title{Write Assessment Items in QTI XML}
\usage{
write_qti(x, file, ...)
}
\arguments{
\item{x}{Object of class \dQuote{\code{qti_item}} to be written.}

\item{file}{Path to file or connection to write to.}

\item{\dots}{Further arguments passed to \code{xml_write}.}
}
\value{
When \code{file} is not provided, x is written to the console
with \code{as.character}.
}
\description{
These functions write QTI XML to files. Only the single and
multiple choice item types are currently supported.
}
\keyword{methods}
