% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_qti_manifest.R
\name{write_qti_manifest}
\alias{write_qti_manifest}
\title{Write QTI XML Manifest and Assessment Package}
\usage{
write_qti_manifest(x, file, ...)
}
\arguments{
\item{x}{\code{qti_manifest} object to be written, or a list of
\code{qti_item} objects to be combined in an assessment package.}

\item{file}{Path to file or connection to write to.}

\item{\dots}{Further arguments passed to \code{xml_write}.}
}
\value{
When \code{file} is not provided, x is written to the console
with \code{as.character}.
}
\description{
These functions write QTI XML manifests to files, and build assessment
packages, with items as input.
}
\keyword{methods}
