% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qti_item.R
\name{qti_item}
\alias{qti_item}
\title{The qti_item class}
\usage{
qti_item(
  id = NULL,
  title = NULL,
  type = c("choice"),
  prompt = NULL,
  options = NULL,
  key = NULL,
  rm_prefix = TRUE,
  href = sprintf("qti-item-\%s.xml", id),
  xml = NULL
)
}
\arguments{
\item{id}{Item ID as an integer or string.}

\item{title}{Item title, as a string, with max 280 characters.}

\item{type}{Item type, only "choice" is currently supported.}

\item{prompt}{Stem text, as a single string.}

\item{options}{Option text, as vector of strings, one per option.}

\item{key}{Numeric vector of 0/1 response scores, one per option.}

\item{rm_prefix}{Logical with default \code{TRUE} for removing prefixes,
e.g., \code{a.}, \code{b.}, \code{c.} from prompt and options.}

\item{href}{Optional file name for item, with default created using
\code{id}.}

\item{xml}{Optional XML representation of item, as character. If not
supplied, this is created automatically from the remaining arguments.}
}
\value{
\code{qti_item} returns an object of class
\dQuote{\code{qti_item}}, which is a list containing named elements for
the item \code{id}, \code{title}, and \code{type}, as supplied via
arguments, the \code{prompt} and \code{options} text with formatting and
resources stripped, and the QTI XML representation of the item.
}
\description{
These functions build and manipulate qti_item objects, for analysis and
for reading and writing to QTI XML.
}
\examples{

item <- qti_item(
  id = 999,
  title = "Example Item",
  type = "choice",
  prompt = "What does this image tell you? <img src='life.png' />",
  options = c("Everything", "Something",
    "Nothing, but look at this code:<br/><pre>lm(life ~ R)</pre>"),
  key = c(1, 1, 0)
)
item

}
\keyword{methods}
