% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qti.R
\name{read_qti}
\alias{read_qti}
\title{Read Assessment Items in QTI XML}
\usage{
read_qti(x, ...)
}
\arguments{
\item{x}{String or connection to QTI XML file.}

\item{\dots}{further arguments passed to or from other functions.}
}
\value{
\code{read_qti} returns an object of class
\dQuote{\code{qti_item}}, which contains the following elements,
if present in x:
\item{id}{item id, extracted from the identifier attribute}
\item{title}{item title, extracted from the title attribute}
\item{type}{item type, currently \dQuote{\code{single}} for single choice}
\item{prompt}{text for the item stem}
\item{options}{list of text for options}
\item{key}{vector of correct/incorrect as 0/1 for options}
\item{xml}{raw XML contents, stored as an object of class
\dQuote{xml_document}}
}
\description{
These functions read QTI XML from character strings or files.
Only the single and multiple choice item types are currently supported.
}
\examples{

# Read from character string
item <- read_qti(
  "<assessmentItem identifier='1584'
    title='Excellent Example Item'>
    <responseDeclaration identifier='RESPONSE' cardinality='single'
      baseType='identifier'>
      <correctResponse>
        <value>ChoiceB</value>
      </correctResponse>
    </responseDeclaration>
    <outcomeDeclaration identifier='SCORE' cardinality='single'
      baseType='float'>
      <defaultValue><value>0</value></defaultValue>
    </outcomeDeclaration>
    <itemBody>
      <prompt>
        <p>Is this a really good question?</p>
      </prompt>
      <choiceInteraction responseIdentifier='RESPONSE' shuffle='false'
        maxChoices='1'>
        <simpleChoice identifier='ChoiceA'>Yes</simpleChoice>
        <simpleChoice identifier='ChoiceB'>No</simpleChoice>
        <simpleChoice identifier='ChoiceC'>Maybe</simpleChoice>
      </choiceInteraction>
    </itemBody>
  </assessmentItem>"
)
item

}
\keyword{methods}
